NeuralAnimator/requirements.txt
tqdm
upc-pymotion
wandb
torch